+++ src/roq/binance_futures/order_entry_simple.cpp
@@ -1329,7 +1329,7 @@ void OrderEntrySimple::process_response(web::rest::Response const &response, aut
         break;
       case SERVER_ERROR: {
         auto message = fmt::format("{}"sv, status);
-        error_handler(Origin::EXCHANGE, RequestStatus::ERROR, Error::UNKNOWN, message);
+        error_handler(Origin::EXCHANGE, RequestStatus::REJECTED, Error::UNKNOWN, message);
         break;
       }
     }