+++ src/roq/binance_futures/rest_trade.cpp
@@ -691,7 +691,7 @@ void RestTrade::process_response(web::rest::Response const &response, auto error
         break;
       case SERVER_ERROR: {
         auto message = fmt::format("{}"sv, status);
-        error_handler(Origin::EXCHANGE, RequestStatus::ERROR, Error::UNKNOWN, message);
+        error_handler(Origin::EXCHANGE, RequestStatus::REJECTED, Error::UNKNOWN, message);
         break;
       }
     }